HOTFIX - Update to classified ad submission email receipt.
UPDATE FOR SHEILA:

"So I am wondering if you could change the following message that people receive on placing their ad:
CLASSIFIED AD  DETAILS
Receipt #1381992051

Thank you for submitting your ad to PennywiseAds.com.
This email is your receipt and should be kept for your records.

CLASSIFIED AD  DETAILS

The line that says:

This email is your receipt and should be kept for your records.

Please Change to:

This email confirms that we have received your ad request. You will receive a follow-up message with an ad number and confirmation of the amount to be charged to your credit card.

Further to this, I delete the following information that appears at the bottom when I send the reply to them so if you could take it off, it would save me that step - note: I leave the Email: info@pennywiseads.com information

DELETE:
Address: Box 430,
         Kaslo BC
         Canada V0G 1M0

PLEASE NOTE:
  Live Load ads are given priority and will be processed as quickly as
possible (during regular business hours).
  Regular ads placed before 3pm on Thursday will appear in the following
Tuesday's print edition.

Thank you for placing your ad with the Pennywise, we appreciate your
business.

Sincerely,
The Pennywise Team.

ALSO: When would we be able to have the Classification choices changed to alphabetical listings on the submission form and when we are putting Live Loads up?
"